+++ CHANGELOG.md
@@ -5,10 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.3.0] - 2020-12-11
 ### Added
 - option `CC1101(lock_spi_device=True)` sets an advisory, exclusive,
   non-blocking `flock` on the SPI device file `/dev/spidev?.?`
-- method `.unlock_spi_device()` to unlock manually
+- method `.unlock_spi_device()` to release `flock` manually
 
 ## [2.2.0] - 2020-12-08
 ### Added
@@ -57,7 +59,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - disabled data whitening
 - automatic calibration
 
-[Unreleased]: https://github.com/fphammerle/python-cc1101/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/fphammerle/python-cc1101/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/fphammerle/python-cc1101/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/fphammerle/python-cc1101/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/fphammerle/python-cc1101/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/fphammerle/python-cc1101/compare/v1.2.0...v2.0.0